Kernel/Net: Move ethernet files to a new directory
