+++ Kernel/CMakeLists.txt
@@ -270,13 +270,13 @@ set(KERNEL_SOURCES
     Library/KLexicalPath.cpp
     Library/KString.cpp
     Library/UserOrKernelBuffer.cpp
-    Net/Intel/E1000ENetworkAdapter.cpp
-    Net/Intel/E1000NetworkAdapter.cpp
-    Net/Realtek/RTL8168NetworkAdapter.cpp
-    Net/VirtIO/VirtIONetworkAdapter.cpp
+    Net/Ethernet/Intel/E1000ENetworkAdapter.cpp
+    Net/Ethernet/Intel/E1000NetworkAdapter.cpp
+    Net/Ethernet/Realtek/RTL8168NetworkAdapter.cpp
+    Net/Ethernet/VirtIO/VirtIONetworkAdapter.cpp
+    Net/Ethernet/Loopback/LoopbackAdapter.cpp
     Net/IPv4/Socket.cpp
     Net/LocalSocket.cpp
-    Net/LoopbackAdapter.cpp
     Net/NetworkAdapter.cpp
     Net/NetworkTask.cpp
     Net/NetworkingManagement.cpp
