+++ Kernel/Net/Ethernet/Intel/E1000ENetworkAdapter.cpp
@@ -7,7 +7,7 @@
 #include <AK/MACAddress.h>
 #include <Kernel/Bus/PCI/API.h>
 #include <Kernel/Bus/PCI/IDs.h>
-#include <Kernel/Net/Intel/E1000ENetworkAdapter.h>
+#include <Kernel/Net/Ethernet/Intel/E1000ENetworkAdapter.h>
 #include <Kernel/Net/NetworkingManagement.h>
 #include <Kernel/Sections.h>
