+++ Kernel/Net/Ethernet/Intel/E1000ENetworkAdapter.h
@@ -11,7 +11,7 @@
 #include <Kernel/Bus/PCI/Device.h>
 #include <Kernel/Interrupts/IRQHandler.h>
 #include <Kernel/Library/IOWindow.h>
-#include <Kernel/Net/Intel/E1000NetworkAdapter.h>
+#include <Kernel/Net/Ethernet/Intel/E1000NetworkAdapter.h>
 #include <Kernel/Net/NetworkAdapter.h>
 #include <Kernel/Security/Random.h>
