+++ Kernel/Net/Ethernet/Intel/E1000NetworkAdapter.cpp
@@ -8,7 +8,7 @@
 #include <Kernel/Bus/PCI/API.h>
 #include <Kernel/Bus/PCI/IDs.h>
 #include <Kernel/Debug.h>
-#include <Kernel/Net/Intel/E1000NetworkAdapter.h>
+#include <Kernel/Net/Ethernet/Intel/E1000NetworkAdapter.h>
 #include <Kernel/Net/NetworkingManagement.h>
 #include <Kernel/Sections.h>
