+++ Kernel/Net/Ethernet/Loopback/LoopbackAdapter.h
