+++ Kernel/Net/NetworkingManagement.cpp
@@ -10,12 +10,12 @@
 #include <Kernel/Bus/PCI/API.h>
 #include <Kernel/Library/KString.h>
 #include <Kernel/Memory/AnonymousVMObject.h>
-#include <Kernel/Net/Intel/E1000ENetworkAdapter.h>
-#include <Kernel/Net/Intel/E1000NetworkAdapter.h>
-#include <Kernel/Net/LoopbackAdapter.h>
+#include <Kernel/Net/Ethernet/Intel/E1000ENetworkAdapter.h>
+#include <Kernel/Net/Ethernet/Intel/E1000NetworkAdapter.h>
+#include <Kernel/Net/Ethernet/Loopback/LoopbackAdapter.h>
 #include <Kernel/Net/NetworkingManagement.h>
-#include <Kernel/Net/Realtek/RTL8168NetworkAdapter.h>
-#include <Kernel/Net/VirtIO/VirtIONetworkAdapter.h>
+#include <Kernel/Net/Ethernet/Realtek/RTL8168NetworkAdapter.h>
+#include <Kernel/Net/Ethernet/VirtIO/VirtIONetworkAdapter.h>
 #include <Kernel/Sections.h>
 
 namespace Kernel {
