+++ Kernel/Net/Routing.cpp
@@ -8,7 +8,7 @@
 #include <AK/Singleton.h>
 #include <Kernel/Debug.h>
 #include <Kernel/Locking/MutexProtected.h>
-#include <Kernel/Net/LoopbackAdapter.h>
+#include <Kernel/Net/Ethernet/Loopback/LoopbackAdapter.h>
 #include <Kernel/Net/NetworkTask.h>
 #include <Kernel/Net/NetworkingManagement.h>
 #include <Kernel/Net/Routing.h>
