+++ Kernel/Net/TCPSocket.h
@@ -15,6 +15,7 @@
 #include <Kernel/Library/LockWeakPtr.h>
 #include <Kernel/Locking/MutexProtected.h>
 #include <Kernel/Net/IPv4/Socket.h>
+#include <Kernel/Net/Routing.h>
 #include <Kernel/Time/TimerQueue.h>
 
 namespace Kernel {
